pyproject.toml
[project]
name = "inteliaudit"
version = "0.1.0"
description = "SaaS de automatización de auditoría impositiva para Paraguay"
requires-python = ">=3.12"
dependencies = [
    # Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    # Base de datos
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    # Config & validación
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    # Scraping
    "playwright>=1.44",
    "httpx>=0.27",
    # Parseo de archivos
    "openpyxl>=3.1",
    "lxml>=5.2",
    "xmltodict>=0.13",
    "pdfplumber>=0.11",
    # Generación de informes
    "jinja2>=3.1",
    "python-docx>=1.1",
    "weasyprint>=62",
    # Autenticación
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "cryptography>=42.0",
    # IA
    "anthropic>=0.28",
    # CLI & terminal
    "click>=8.1",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
inteliaudit = "main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["api", "db", "analisis", "ingesta", "informes", "papeles", "config", "landing", "ui"]
